add cipher function
