+++ cipher.js
@@ -1,20 +1,16 @@
-const alpahbet_array = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z']
-const shift = (string, key) => {
-  const indexOfString = alpahbet_array.indexOf(string);
-  let newString
-  let newIndex = indexOfString + key
-  if (newIndex > 25 && newIndex < 52) {
-    newIndex = (newIndex - 25) - 1
-    newString = alpahbet_array[newIndex];
-    return newString
-  } else {
-    newIndex = indexOfString + key
-    newString = alpahbet_array[newIndex];
-    return newString
+import shift from './shift';
+
+const cipher = (string, key) => {
+  let stripped = string.split("");
+  let empty = []
+  for (let i = 0; i < stripped.length; i++) {
+    empty.push(shift(stripped[i], key))
+    empty
   }
+  let myArr = empty
+  myArr = [myArr.join('')]
+  var finalString = myArr.toString()
+  return finalString
 }
 
-export default shift;
-
-//efgfoe uif fbtu xbmm pg uif dbtumf
-//defend the east wall of the castle
+export default cipher;
