+++ cipher.test.js
@@ -1,5 +1,10 @@
-import shift from './cipher';
+import shift from './shift';
+import cipher from './cipher';
 
 test('shifts one letter', () => {
   expect(shift('z', 3)).toBe('c');
+});
+
+test('ciphers strings', () => {
+  expect(cipher('abcdefghijklmnopqrstuvwxyz', 1)).toBe('bcdefghijklmnopqrstuvwxyza');
 });
